+++ web/packages/gallery/services/magic-metadata.ts
